add docstrings for sga
