+++ AOEMods.Essence.Editor/ArchiveViewModel.cs
@@ -52,8 +52,7 @@ namespace AOEMods.Essence.Editor
                 if (saveFileDialog.ShowDialog() == true)
                 {
                     using var outFile = File.Open(saveFileDialog.FileName, FileMode.Create, FileAccess.Write);
-                    var archiveWriter = new ArchiveWriter(outFile);
-                    archiveWriter.Write(Archive);
+                    ArchiveWriterHelper.WriteArchiveToStream(outFile, Archive);
                 }
             }
         }
