+++ AOEMods.Essence/SGA/Core/ArchiveWriter.cs
@@ -3,15 +3,11 @@ using System.Text;
 
 namespace AOEMods.Essence.SGA.Core;
 
+/// <summary>
+/// Extends BinaryWriter to provide functions for writing SGA archive files.
+/// </summary>
 public class ArchiveWriter : BinaryWriter
 {
-    private static readonly byte[] Magic = new byte[] { 95, 65, 82, 67, 72, 73, 86, 69 };
-    private const ushort Version = 10;
-    private const ushort Product = 0;
-    private const uint BlockSize = 262144;
-    private const ulong HeaderSize = 428UL;
-    private const uint HeaderExtraSize = 44;
-
     public ArchiveWriter(Stream output) : base(output)
     {
     }
@@ -31,6 +27,12 @@ public class ArchiveWriter : BinaryWriter
     private MemoryStream archiveData = new();
     private MemoryStream archiveStrings = new();
 
+    /// <summary>
+    /// Adds a string to the archive's string blob. Use WriteStrings to write the
+    /// string blob to the stream.
+    /// </summary>
+    /// <param name="s">String to add to the archive's string blob.</param>
+    /// <returns>Offset within the string blob where the added string starts.</returns>
     public long AddString(string s)
     {
         long stringOffset = archiveStrings.Length;
@@ -39,6 +41,12 @@ public class ArchiveWriter : BinaryWriter
         return stringOffset;
     }
 
+    /// <summary>
+    /// Adds a string to the archive's data blob. Use WriteData to write the
+    /// data blob to the stream.
+    /// </summary>
+    /// <param name="data">Data to add to the archive's data blob.</param>
+    /// <returns>Offset within the data blob where the added data starts.</returns>
     public long AddData(byte[] data)
     {
         long offset = archiveData.Length;
@@ -46,17 +54,25 @@ public class ArchiveWriter : BinaryWriter
         return offset;
     }
 
+    /// <summary>
+    /// Writes the string blob previously created using AddString
+    /// to the stream and advances the stream's position.
+    /// </summary>
     public void WriteStrings()
     {
         Write(archiveStrings.ToArray());
     }
 
+    /// <summary>
+    /// Writes the data blob previously created using AddData
+    /// to the stream and advances the stream's position.
+    /// </summary>
     public void WriteData()
     {
         Write(archiveData.ToArray());
     }
 
-    public void WriteFixedString(string s, int charCount, int charSize)
+    private void WriteFixedString(string s, int charCount, int charSize)
     {
         var nameBytes = charSize switch
         {
@@ -68,25 +84,30 @@ public class ArchiveWriter : BinaryWriter
         Write(new byte[charCount * charSize - nameBytes.Length]);
     }
 
+    /// <summary>
+    /// Writes an archive header to the stream and advances the stream's position.
+    /// </summary>
+    /// <param name="header">Archive header to write to the stream.</param>
+    /// <exception cref="Exception">Thrown if the header length or offset is inconsistent.</exception>
     public void Write(ArchiveHeader header)
     {
         Write(header.Magic); // 0
         Write(header.Version); // 8
         Write(header.Product); // 10
         WriteFixedString(header.NiceName, 64, 2); // 12
-        Write(header.Offset); // 140 header blob offset
+        Write(header.HeaderBlobOffset); // 140 header blob offset
         Write(header.HeaderBlobLength); // 148 num2 / header blob length
         Write(header.DataOffset); // 152 data blob offset
         Write(header.DataBlobLength); // 160 data blob length
         Write(1U); // 168 unk2 == 1
 
         Write(header.Signature);
-        if (header.Offset != (ulong)BaseStream.Position)
+        if (header.HeaderBlobOffset != (ulong)BaseStream.Position)
         {
             throw new Exception("Header length or offset incorrect.");
         }
 
-        Write(new byte[header.Offset - (ulong)BaseStream.Position]);
+        Write(new byte[header.HeaderBlobOffset - (ulong)BaseStream.Position]);
 
         Write(header.TocDataOffset); // 256
         Write(header.TocDataCount); // 260
@@ -103,6 +124,10 @@ public class ArchiveWriter : BinaryWriter
         Write(header.BlockSize); // 296
     }
 
+    /// <summary>
+    /// Writes an archive table of contents entry to the stream and advances the stream's position.
+    /// </summary>
+    /// <param name="tocEntry">Archive table of contents entry to write to the stream.</param>
     public void Write(ArchiveTocEntry tocEntry)
     {
         WriteFixedString(tocEntry.Alias, 64, 1);
@@ -114,6 +139,10 @@ public class ArchiveWriter : BinaryWriter
         Write(tocEntry.FolderRootIndex);
     }
 
+    /// <summary>
+    /// Writes an archive folder entry to the stream and advances the stream's position.
+    /// </summary>
+    /// <param name="folderEntry">Archive folder entry to write to the stream.</param>
     public void Write(ArchiveFolderEntry folderEntry)
     {
         Write(folderEntry.NameOffset);
@@ -123,6 +152,10 @@ public class ArchiveWriter : BinaryWriter
         Write(folderEntry.FileEndIndex);
     }
 
+    /// <summary>
+    /// Writes an archive file entry to the stream and advances the stream's position.
+    /// </summary>
+    /// <param name="fileEntry">Archive file entry to write to the stream.</param>
     public void Write(ArchiveFileEntry fileEntry)
     {
         Write(fileEntry.NameOffset);
@@ -133,96 +166,5 @@ public class ArchiveWriter : BinaryWriter
         Write((byte)fileEntry.VerificationType);
         Write((byte)fileEntry.StorageType);
         Write(fileEntry.Crc);
-    }
-
-    public void Write(IArchive archive)
-    {
-        var toc = archive.Tocs[0];
-        toc.RebuildFromRootFolder();
-
-        var rootFolder = toc.RootFolder;
-        var fileNodes = toc.Files;
-        var folderNodes = toc.Folders;
-
-        MemoryStream contentStream = new();
-        ArchiveWriter writer = new(contentStream, Encoding.ASCII);
-
-        var fileDataOffsets = fileNodes.Select(fileNode =>
-        {
-            var fileData = fileNode.GetData().ToArray();
-            return (writer.AddData(fileData), fileData.Length);
-        }).ToArray();
-
-        var directoryNameOffsets = folderNodes.Select(node => writer.AddString(node.FullName)).ToArray();
-        var fileNameOffsets = fileNodes.Select(file => writer.AddString(file.Name)).ToArray();
-
-        // Write files
-        long fileEntryOffset = writer.BaseStream.Position;
-        for (int i = 0; i < fileNodes.Count; i++)
-        {
-            ulong fileDataOffset = (ulong)fileDataOffsets[i].Item1;
-            uint fileDataLength = (uint)fileDataOffsets[i].Item2;
-            uint fileNameOffset = (uint)fileNameOffsets[i];
-
-            writer.Write(new ArchiveFileEntry(
-                fileNameOffset, 0, fileDataOffset,
-                fileDataLength, fileDataLength,
-                0, 0, 0
-            ));
-        }
-
-        // Write folders
-        long folderEntryOffset = writer.BaseStream.Position;
-        for (int i = 0; i < folderNodes.Count; i++)
-        {
-            var dir = folderNodes[i];
-            var dirFiles = dir.Children.OfType<IArchiveFileNode>().ToArray();
-            var dirDirectories = dir.Children.OfType<IArchiveFolderNode>().ToArray();
-            uint dirNameOffset = (uint)directoryNameOffsets[i];
-
-            writer.Write(new ArchiveFolderEntry(
-                dirNameOffset,
-                dirDirectories.Length == 0 ? 0 : (uint)folderNodes.IndexOf(dirDirectories.First()),
-                dirDirectories.Length == 0 ? 0 : (uint)folderNodes.IndexOf(dirDirectories.Last()) + 1,
-                dirFiles.Length == 0 ? 0 : (uint)fileNodes.IndexOf(dirFiles.First()),
-                dirFiles.Length == 0 ? 0 : (uint)fileNodes.IndexOf(dirFiles.Last()) + 1
-            ));
-        }
-
-        // Write toc
-        long tocEntryOffset = writer.BaseStream.Position;
-
-        writer.Write(new ArchiveTocEntry(
-            toc.Alias, toc.Name,
-            0, (uint)toc.Files.Count,
-            0, (uint)toc.Folders.Count,
-            (uint)folderNodes.IndexOf(toc.RootFolder)
-        ));
-
-        long stringsOffset = writer.BaseStream.Position;
-        writer.WriteStrings();
-        long stringLength = writer.BaseStream.Position - stringsOffset;
-
-        long headerBlobLength = writer.BaseStream.Position + 44;
-
-        long dataOffset = writer.BaseStream.Position;
-        writer.WriteData();
-        long dataBlobLength = writer.BaseStream.Position - dataOffset;
-
-        // Write the actual file
-
-        // Write header
-        Write(new ArchiveHeader(
-            Magic, Version, Product, archive.Name,
-            HeaderSize, (uint)headerBlobLength,
-            HeaderSize + (ulong)dataOffset + HeaderExtraSize, (ulong)dataBlobLength,
-            (uint)tocEntryOffset + HeaderExtraSize, 1,
-            (uint)folderEntryOffset + HeaderExtraSize, (uint)folderNodes.Count,
-            (uint)fileEntryOffset + HeaderExtraSize, (uint)fileNodes.Count,
-            (uint)stringsOffset + HeaderExtraSize, (uint)stringLength,
-            BlockSize, archive.Signature
-        ));
-
-        Write(contentStream.ToArray());
     }
 }
