+++ AOEMods.Essence/SGA/Graph/ArchiveFileNode.cs
@@ -4,9 +4,19 @@ using System.Text;
 
 namespace AOEMods.Essence.SGA.Graph;
 
+/// <summary>
+/// File node of an SGA archive containing data read from a stream as stored in an SGA file.
+/// </summary>
 public class ArchiveFileNode : IArchiveFileNode
 {
+    /// <summary>
+    /// Parent of the node.
+    /// </summary>
     public IArchiveNode? Parent { get; }
+
+    /// <summary>
+    /// Name of the node.
+    /// </summary>
     public string Name { get; set; }
 
     private Stream dataStream;
@@ -15,6 +25,16 @@ public class ArchiveFileNode : IArchiveFileNode
     private long dataUncompressedLength;
     private FileStorageType storageType;
 
+    /// <summary>
+    /// Initializes an ArchiveFileNode from values as usually stored in an SGA file.
+    /// </summary>
+    /// <param name="dataStream">Stream to read the node's data from.</param>
+    /// <param name="name">Name of the node.</param>
+    /// <param name="dataPosition">Position of the data in the stream.</param>
+    /// <param name="dataLength">Size of the data in bytes in the stream.</param>
+    /// <param name="dataUncompressedLength">Size of the data when uncompressed.</param>
+    /// <param name="storageType">How the data is stored in the stream.</param>
+    /// <param name="parent">Parent of the node.</param>
     public ArchiveFileNode(Stream dataStream, string name, long dataPosition, long dataLength, long dataUncompressedLength, FileStorageType storageType, IArchiveNode? parent = null)
     {
         Name = name;
@@ -26,6 +46,10 @@ public class ArchiveFileNode : IArchiveFileNode
         this.storageType = storageType;
     }
 
+    /// <summary>
+    /// Reads and returns the data of the file node from the stream.
+    /// </summary>
+    /// <returns>Data of the file.</returns>
     public IEnumerable<byte> GetData()
     {
         dataStream.Position = dataPosition;
@@ -33,7 +57,7 @@ public class ArchiveFileNode : IArchiveFileNode
         switch (storageType)
         {
             case FileStorageType.Store:
-                BinaryReader reader = new BinaryReader(dataStream, Encoding.ASCII, true);
+                BinaryReader reader = new BinaryReader(dataStream, Encoding.UTF8, true);
                 return reader.ReadBytes((int)dataUncompressedLength);
             case FileStorageType.StreamCompress:
             case FileStorageType.BufferCompress:
